+++ src/main/java/com/solegendary/reignofnether/unit/units/villagers/RavagerUnit.java
@@ -314,7 +314,7 @@ public class RavagerUnit extends Ravager implements Unit, AttackerUnit {
                     }
                 }
                 for (BuildingPlacement building : affectedBuildings)
-                    building.destroyRandomBlocks((int) ROAR_DAMAGE);
+                    building.destroyRandomBlocks(ROAR_DAMAGE);
 
                 Vec3 vec3 = this.getBoundingBox().getCenter();
