+++ src/main/java/com/solegendary/reignofnether/unit/units/villagers/RoyalGuardUnit.java
@@ -625,7 +625,7 @@ public class RoyalGuardUnit extends Vindicator implements AttackerUnit, HeroUnit
                 }
             }
             for (BuildingPlacement building : buildings) {
-                building.destroyRandomBlocks((int) (maceSlam.damage / 2));
+                building.destroyRandomBlocks(maceSlam.damage / 2);
             }
 
             for (LivingEntity hitEntity : MiscUtil.getEntitiesWithinRange(Vec3.atCenterOf(blockPos.above()), MaceSlam.RADIUS, LivingEntity.class, level())) {
